Networking 클래스를 별도의 파일로 모듈화
script.js 에 별도로 분리된 Networking.js 내 클래스를 사용할 수 있도록 모듈화
